feat: add delete button to edit countdown page and remove delete button from countdown page
