+++ src/modules/countdown/ui/sections/countdown-header-section.tsx
@@ -2,14 +2,9 @@
 
 import { Button } from "@/components/ui/button";
 import { useTRPC } from "@/trpc/react";
-import {
-  useMutation,
-  useQueryClient,
-  useSuspenseQuery,
-} from "@tanstack/react-query";
-import { Edit, Trash2 } from "lucide-react";
+import { useSuspenseQuery } from "@tanstack/react-query";
+import { Edit } from "lucide-react";
 import Link from "next/link";
-import { useRouter } from "next/navigation";
 import { Suspense } from "react";
 import { ErrorBoundary } from "react-error-boundary";
 
@@ -36,42 +31,13 @@ const CountdownHeaderSectionSkeleton = () => {
 const CountdownHeaderSectionSuspense = ({
   countdownId,
 }: CountdownHeaderSectionProps) => {
-  const router = useRouter();
   const trpc = useTRPC();
-  const queryClient = useQueryClient();
 
   const { data: countdown } = useSuspenseQuery({
     ...trpc.countdown.getById.queryOptions({
       id: parseInt(countdownId),
     }),
   });
-
-  const deleteCountdownMutation = useMutation({
-    ...trpc.countdown.delete.mutationOptions(),
-    onSuccess: () => {
-      void queryClient.invalidateQueries({
-        queryKey: trpc.countdown.getAll.queryKey(),
-      });
-
-      void router.push("/dashboard");
-    },
-    onError: (error) => {
-      console.error("Failed to delete countdown:", error);
-      // Maybe add toast notification here if you have a toast system
-    },
-  });
-
-  const handleDelete = () => {
-    if (
-      confirm(
-        `Are you sure you want to delete "${countdown.name}"? This action cannot be undone.`,
-      )
-    ) {
-      deleteCountdownMutation.mutate({
-        id: countdown.id,
-      });
-    }
-  };
 
   return (
     <section className="py-8 md:py-12">
@@ -81,23 +47,13 @@ const CountdownHeaderSectionSuspense = ({
             {countdown.name}
           </h1>
         </div>
-        <div className="flex gap-2">
-          <Button asChild variant="outline" size="sm">
-            <Link href={`/countdown/${countdown.id}/edit`}>
-              <Edit className="h-4 w-4" />
-              Edit
-            </Link>
-          </Button>
-          <Button
-            variant="outline"
-            size="sm"
-            onClick={handleDelete}
-            disabled={deleteCountdownMutation.isPending}
-          >
-            <Trash2 className="h-4 w-4" />
-            {deleteCountdownMutation.isPending ? "Deleting..." : "Delete"}
-          </Button>
-        </div>
+
+        <Button asChild variant="outline" size="sm">
+          <Link href={`/countdown/${countdown.id}/edit`}>
+            <Edit className="h-4 w-4" />
+            Edit
+          </Link>
+        </Button>
       </div>
     </section>
   );
